requirements.txt
streamlit
openai
python-docx
pandas
matplotlib
